Tidy up newsletter and consent datashaping
